Drop TopicRetriever#perform_retrieve monkey patch
`TopicRetriever` is monkey-patched so that a topic could be fetched
directly from its feed when `/embed/comments` is hit.

However, since feed polling is rate limited, the monkey patched method
will not ran in most cases and the post will be eventually updated when
the next time the job runs.

This closes #2.