DEV: Improve admin interface UI
The rss feeds used to be stored in a single site setting. They are now
stored in the a proper db table, but the admin interface was never
updated to reflect this. This commit updates the admin interface UI so
that there are individual create, edit, and destroy buttons for each
feed.

This should also address: https://meta.discourse.org/t/343372

Follow up to: 73ac61823cb663295fa673d4488acbc6822c4711